Switch to uint64

+++ libtiledbsoma/src/soma/column_buffer.h
@@ -346,17 +346,17 @@ class ColumnBuffer {
     /**
      * @brief Resize the internal buffers to the given size.
      */
-    void resize(size_t size, size_t num_cells, bool preserve_data = false);
+    void resize(uint64_t size, uint64_t num_cells, bool preserve_data = false);
 
     /**
      * @brief Get max size of data buffer in bytes.
      */
-    size_t max_size() const;
+    uint64_t max_size() const;
 
     /**
      * @brief Get the max number of cells the buffers can hold.
      */
-    size_t max_num_cells() const;
+    uint64_t max_num_cells() const;
 
    private:
     //===================================================================
